Renamed SimplifiedInnerRetrieverUtils to MultiFieldsInnerRetrieverUtils
